+++ common/src/test/java/com/microsoft/identity/common/internal/telemetry/OnboardingSessionCorrelationStoreTest.java
@@ -1,0 +1,69 @@
+// Copyright (c) Microsoft Corporation.
+// All rights reserved.
+//
+// This code is licensed under the MIT License.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files(the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions :
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+package com.microsoft.identity.common.internal.telemetry;
+
+import androidx.test.core.app.ApplicationProvider;
+
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.robolectric.RobolectricTestRunner;
+
+@RunWith(RobolectricTestRunner.class)
+public class OnboardingSessionCorrelationStoreTest {
+
+    private OnboardingSessionCorrelationStore mPersistence;
+
+    @Before
+    public void setup() {
+        mPersistence = new OnboardingSessionCorrelationStore(
+                ApplicationProvider.getApplicationContext());
+    }
+
+    @Test
+    public void testLoad_Empty_ReturnsEmptyString() {
+        Assert.assertEquals("", mPersistence.load());
+    }
+
+    @Test
+    public void testSaveAndLoad() {
+        final String json = "{\"key|scope\":{\"id\":\"uuid\",\"ts\":1234567890}}";
+        mPersistence.save(json);
+        Assert.assertEquals(json, mPersistence.load());
+    }
+
+    @Test
+    public void testSave_OverwritesPrevious() {
+        mPersistence.save("first");
+        mPersistence.save("second");
+        Assert.assertEquals("second", mPersistence.load());
+    }
+
+    @Test
+    public void testSave_EmptyString() {
+        mPersistence.save("some data");
+        mPersistence.save("");
+        Assert.assertEquals("", mPersistence.load());
+    }
+}
