+++ common/src/test/java/com/microsoft/identity/common/internal/telemetry/OnboardingTelemetryRecorderTest.java
@@ -1,10 +1,30 @@
 // Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
+// All rights reserved.
+//
+// This code is licensed under the MIT License.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files(the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions :
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
 package com.microsoft.identity.common.internal.telemetry;
 
 import androidx.test.core.app.ApplicationProvider;
 
-import com.microsoft.identity.common.java.telemetry.OnboardingBlobFieldKeys;
+import com.microsoft.identity.common.java.telemetry.OnboardingTelemetryConstants;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
@@ -107,21 +127,23 @@ public class OnboardingTelemetryRecorderTest {
     @Test
     public void testAddStep_AppearsInFinalizedBlob() throws Exception {
         mRecorder.addBlockingError("BROKER_INSTALLATION_TRIGGERED");
-        mRecorder.addStep(OnboardingBlobFieldKeys.STEP_AUTHENTICATION_STARTED, "2025-10-29T15:00:00Z");
-        mRecorder.addStep(OnboardingBlobFieldKeys.STEP_BROKER_INSTALL_PROMPTED, "2025-10-29T15:00:05Z");
+        mRecorder.addStep(OnboardingTelemetryConstants.STEP_AUTHENTICATION_STARTED);
+        mRecorder.addStep(OnboardingTelemetryConstants.STEP_BROKER_INSTALL_PROMPTED);
 
         final JSONObject blob = new JSONObject(mRecorder.finalizeBlob());
         final JSONArray steps = blob.getJSONArray("steps_list");
         Assert.assertEquals(2, steps.length());
         Assert.assertEquals("AuthenticationStarted", steps.getJSONObject(0).getString("step_id"));
+        Assert.assertTrue(steps.getJSONObject(0).has("ts"));
         Assert.assertEquals("BrokerInstallPrompted", steps.getJSONObject(1).getString("step_id"));
+        Assert.assertTrue(steps.getJSONObject(1).has("ts"));
     }
 
     @Test
     public void testLastCompletedStep_SetAutomatically() throws Exception {
         mRecorder.addBlockingError("BROKER_INSTALLATION_TRIGGERED");
-        mRecorder.addStep(OnboardingBlobFieldKeys.STEP_AUTHENTICATION_STARTED, "2025-10-29T15:00:00Z");
-        mRecorder.addStep(OnboardingBlobFieldKeys.STEP_BROKER_INSTALL_PROMPTED, "2025-10-29T15:00:01Z");
+        mRecorder.addStep(OnboardingTelemetryConstants.STEP_AUTHENTICATION_STARTED);
+        mRecorder.addStep(OnboardingTelemetryConstants.STEP_BROKER_INSTALL_PROMPTED);
 
         final JSONObject blob = new JSONObject(mRecorder.finalizeBlob());
         Assert.assertEquals("BrokerInstallPrompted", blob.getString("last_completed_step"));
@@ -151,7 +173,7 @@ public class OnboardingTelemetryRecorderTest {
     @Test
     public void testSetProfile() throws Exception {
         mRecorder.addBlockingError("BROKER_INSTALLATION_TRIGGERED");
-        mRecorder.setProfile(OnboardingBlobFieldKeys.PROFILE_WORK);
+        mRecorder.setProfile(OnboardingTelemetryConstants.PROFILE_WORK);
 
         final JSONObject blob = new JSONObject(mRecorder.finalizeBlob());
         Assert.assertEquals("workProfile", blob.getString("profile"));
