+++ common4j/src/main/com/microsoft/identity/common/java/telemetry/OnboardingTelemetryConstants.java
@@ -1,5 +1,25 @@
 // Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
+// All rights reserved.
+//
+// This code is licensed under the MIT License.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files(the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions :
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
 
 package com.microsoft.identity.common.java.telemetry;
 
@@ -9,7 +29,7 @@ package com.microsoft.identity.common.java.telemetry;
  * to produce the final MATS column name (e.g., "blocking_errors" → "mo_blocking_errors").
  * Seed creation + aggregation keys come from OnboardingBlobConstants (Djinni-generated).
  */
-public final class OnboardingBlobFieldKeys {
+public final class OnboardingTelemetryConstants {
     // Field keys for populated blob (written by OnboardingTelemetryRecorder, read by EntityStore with mo_ prefix)
     public static final String BLOCKING_ERRORS = "blocking_errors";
     public static final String LAST_BLOCKING_ERROR = "last_blocking_error";
@@ -35,5 +55,5 @@ public final class OnboardingBlobFieldKeys {
     public static final String PROFILE_USER = "userProfile";
     public static final String PROFILE_WORK = "workProfile";
 
-    private OnboardingBlobFieldKeys() {} // non-instantiable
+    private OnboardingTelemetryConstants() {} // non-instantiable
 }
